refactor(xmlrpc): extract mod tests to sibling file
